feat: VESA/GOP framebuffer
- QEMU Console 창에 출력을 위한 framebuffer를 0x0000'0000'C000'0000에 추가함
- 사실 내가 새로 추가한 게 아니고 이미 GRUB이 준비해 둔 걸 이제서야 쓰는 것뿐
- 일단 kmain에서 앞으로 fb를 사용하기 위해 초기화하는 framebuffer_init을 만들었고 거기서 아직 페이지 테이블에 등록되지 않은 프레임버퍼의 영역을 페이지 테이블에 2MB 단위로 등록함
- 그리고 바탕색을 단색으로 칠하는 함수인 framebuffer_clear() 함수를 만들어서 바탕색을 흰색으로 출력하는 것까진 완료함

+++ src/mm/slab.c
@@ -4,6 +4,7 @@
 #include <mm/pmm.h>
 #include <mm/slab.h>
 #include <mm/vmm.h>
+#include <stddef.h>
 
 kmem_cache_t kmem_cache_list[KMEM_CACHE_COUNT];	// KMEM_CACHE_COUNT는 9 (mm/slab.h)
 
@@ -128,7 +129,7 @@ void* kmalloc(uint64_t size) {
 
 	int i = 0;
 	for (i = 0; i < KMEM_CACHE_COUNT; i++) {
-		if (size <= (8 << i)) break;
+		if (size <= (8ULL << i)) break;
 	}
 
 	// 8B ~ 2KB 담당
@@ -160,7 +161,7 @@ void slab_init(kmem_cache_t* kmem_cache, slab_t* slab) {
 
 	void* prev = slab->free_list;
 
-	for (int i = 0; i < slab->kmem_cache->obj_capacity - 1; i++) {
+	for (uint64_t i = 0; i < slab->kmem_cache->obj_capacity - 1; i++) {
 		*((uintptr_t*)prev) = (uintptr_t)((uint8_t*)prev + slab->kmem_cache->obj_size);
 		prev = (void*)*((uintptr_t*)prev);
 	}
